+++ src/main/java/database/service/impl/Neo4JEmbeddedServiceImpl.java
@@ -32,7 +32,7 @@ public class Neo4JEmbeddedServiceImpl implements DatabaseService
 	}
 
 	@Override
-	public long getNextAutoIncrement()
+	public int getNextAutoIncrement()
 	{
 		Node autoIncrement = null;
 		try(Transaction transaction = this.graphDatabaseService.beginTx())
@@ -67,11 +67,11 @@ public class Neo4JEmbeddedServiceImpl implements DatabaseService
 		{
 			try(Transaction transaction = this.graphDatabaseService.beginTx())
 			{
-				Integer nextAutoIncrement = (Integer) autoIncrement.getProperty("next");
+				int nextAutoIncrement = (int) autoIncrement.getProperty("next");
 				autoIncrement.setProperty("next", nextAutoIncrement + 1);
 				
 				transaction.success();
-				return nextAutoIncrement.longValue();
+				return nextAutoIncrement;
 			}
 			catch(Exception exception)
 			{
